Bump version 3.0.0
There are some breaking changes, so the next number is 3.0.0.

More information about this changes can be found in the changelog.

The version numbers 2.4.0 and 2.3.2 can be found in the Git history.
These are temporary numbers with no meaning. They don't have a
git tag and certainly don’t have a release on PyPI.