Fix check response for stopped units and units selected with --unit
It has been reported that stopped units do not appear in the command output
of `systemctl list-units --all`
https://github.com/Josef-Friedrich/check_systemd/issues/34#issuecomment-1966675030

It is not enough to parse the command `systemctl list-units --all`. In addition,
the command `systemctl show <unit>` must be called by check_systemd.py,
if the option `--unit` is specified.

See #34