contrib/icinga2: harmonize Icinga 2 config w/ ITL
While fixing the outdated command name in the Icinga 2 ITL[0], it was
suggested to support the newer command line options there as well.

To synchronize or harmonize the ITL definition and the CheckCommand
shipped here, I made small changes to not break the ITL, as I would
guess that more people are using the ITL definition instead of the
shipped command.conf from here.

The breaking changes were added to the CHANGELOG.

[0]: https://github.com/Icinga/icinga2/pull/10035